# phonebook-app/docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/phonebooks/{phonebookId}/update": {
      "put": {
        "tags": [
          "phonebook-controller"
        ],
        "operationId": "put",
        "parameters": [
          {
            "name": "phonebookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePhonebookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDTO"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UpdatePhonebookResponse"
                }
              }
            }
          }
        }
      }
    },
    "/phonebooks/create": {
      "post": {
        "tags": [
          "phonebook-controller"
        ],
        "operationId": "post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePhonebookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDTO"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePhonebookResponse"
                }
              }
            }
          }
        }
      }
    },
    "/phonebooks/search": {
      "get": {
        "tags": [
          "phonebook-controller"
        ],
        "operationId": "get",
        "parameters": [
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortDirection",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDTO"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationSearchPhonebookResponse"
                }
              }
            }
          }
        }
      }
    },
    "/phonebooks/{phonebookId}/delete": {
      "delete": {
        "tags": [
          "phonebook-controller"
        ],
        "operationId": "delete",
        "parameters": [
          {
            "name": "phonebookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDTO"
                }
              }
            }
          },
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorDTO": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "cause": {
            "type": "object"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "UpdatePhonebookRequest": {
        "type": "object",
        "properties": {
          "firstName": {
            "pattern": "^[a-zA-Z ]*$",
            "type": "string"
          },
          "lastName": {
            "pattern": "^[a-zA-Z ]*$",
            "type": "string"
          },
          "phoneNumber": {
            "pattern": "^[0-9]*$",
            "type": "string"
          }
        }
      },
      "UpdatePhonebookResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          }
        }
      },
      "CreatePhonebookRequest": {
        "required": [
          "firstName",
          "lastName",
          "phoneNumber"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "pattern": "^[a-zA-Z ]*$",
            "type": "string"
          },
          "lastName": {
            "pattern": "^[a-zA-Z ]*$",
            "type": "string"
          },
          "phoneNumber": {
            "pattern": "^[0-9]*$",
            "type": "string"
          }
        }
      },
      "CreatePhonebookResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          }
        }
      },
      "PaginationSearchPhonebookResponse": {
        "type": "object",
        "properties": {
          "paging": {
            "$ref": "#/components/schemas/Paging"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchPhonebookResponse"
            }
          }
        }
      },
      "Paging": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "limit": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "SearchPhonebookResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          }
        }
      }
    }
  }
}